% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAT_MH_A_THETA_LAMBDA_FUN_10.R
\name{log_prob_A_lambda_theta_fun}
\alias{log_prob_A_lambda_theta_fun}
\title{Proposal Log Probability of A — Internal}
\usage{
log_prob_A_lambda_theta_fun(
  R,
  W_samp,
  lambda,
  A,
  dict,
  theta_0_samp,
  theta,
  rep_mt,
  col_name,
  motif_len_w
)
}
\arguments{
\item{R}{A character vector of sequences.}

\item{W_samp}{Motif assignments.}

\item{lambda}{Lambda vector for the motif.}

\item{A}{A list of motif position matrices.}

\item{dict}{Alphabet vector.}

\item{theta_0_samp}{Background model.}

\item{theta}{List of motif PWMs.}

\item{rep_mt}{Integer. Index of motif.}

\item{col_name}{Integer. Column index (i.e., motif position) being proposed.}
}
\value{
Log proposal probability.
}
\description{
Computes the log probability of a proposed A (motif positions) under a uniform
proposal distribution for the selected column.
}
\keyword{internal}
