% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAT_LOGLLK_FUN.R
\name{logllk_fun_k}
\alias{logllk_fun_k}
\title{Log-Likelihood under Motif Model — Internal}
\usage{
logllk_fun_k(Ri, wi, ai, theta_0_samp, theta_samp, dict)
}
\arguments{
\item{Ri}{A character vector representing a single sequence.}

\item{wi}{Integer. The motif index assigned to the sequence.}

\item{ai}{A numeric vector of motif start positions.}

\item{theta_0_samp}{A vector of background letter probabilities.}

\item{theta_samp}{A list of motif position weight matrices.}

\item{dict}{A character vector of the sequence alphabet.}
}
\value{
A numeric value representing the log-likelihood.
}
\description{
Computes the log-likelihood of a sequence under a motif model, given the motif
assignment and motif position indices.
}
\keyword{internal}
