% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAT_THETA_FUN.R
\name{theta_samp_fun}
\alias{theta_samp_fun}
\title{Sample Motif Position Weight Matrix Theta — Internal}
\usage{
theta_samp_fun(R, W_samp, A_samp, dict, pri_alw, rep_mt, motif_len_w)
}
\arguments{
\item{R}{A character vector of sequences.}

\item{W_samp}{A numeric vector of motif assignments.}

\item{A_samp}{A list of motif position matrices.}

\item{dict}{A character vector of the sequence alphabet.}

\item{pri_alw}{A numeric vector of prior Dirichlet concentration parameters (motif-wise).}

\item{rep_mt}{Integer. Index of the motif to sample theta for.}
}
\value{
A numeric matrix of sampled motif PWM for the selected motif.
}
\description{
Samples the position weight matrix (PWM) \code{theta} for a specific motif,
using a Dirichlet posterior based on observed positions in assigned sequences.
}
\keyword{internal}
